+++ hpc/nvidia/cuda/cuda_adapter.h
@@ -80,6 +80,9 @@
  */
 #define cudaMalloc(pp, sz)	__cuda(Malloc(pp, sz))
 #define cudaFree(ptr)	__cuda(Free(ptr))
+/**
+ * Pinned memory, on CPU memory, GPU accessable, DMA copy
+ */
 #define cudaMallocHost(ptr, sz, flags)	__cuda(MallocHost(ptr, sz, flags))
 #define cudaFreeHost(ptr, sz, flags)	__cuda(FreeHost(ptr, sz, flags))
 #define cudaMemset(ptr, v, size)	__cuda(Memset(ptr, v, size))